David is irrationally annoyed about documentation standards
Documented floodfill despite my grievances.

+++ brain/src/segmentation.rs
@@ -1,9 +1,19 @@
 use nalgebra as na;
 use std::collections::VecDeque;
 
-const THRESHOLD: u32 = 10;
+const THRESHOLD: f32 = 10.0;
 
-fn floodfill(image: na::SMatrix<na::Vector3<u32>, 1920, 1080>, pos: na::Vector2<usize>) {
+// Returns a color-based floodfill mask with the same dimensions as the input.
+//
+// # Arguments
+// * `image` - A 1920x1080 matrix of Vector3<f32>, where each vector is RGB values from 0-255.
+// # Example
+// ```
+// use nalgebra as na;
+// let image = na::Smatrix::repeat(na::Vector3::new_random()) // Just a demo - not actually RGB
+// let mask = floodfill(image)
+// ```
+fn floodfill(image: na::SMatrix<na::Vector3<f32>, 1920, 1080>, pos: na::Vector2<usize>) {
     let mut q = VecDeque::new();
     q.push_front(image[(pos.x, pos.y)]);
     while !q.is_empty() {
